feat(announcements): restructure announcement views and layouts for better organization
