+++ lib/core/utils/router/app_router.dart
@@ -9,7 +9,7 @@ import 'package:sams_app/core/utils/router/routes_name.dart';
 import 'package:sams_app/core/utils/services/service_locator.dart';
 import 'package:sams_app/core/widgets/shared/general_error_page.dart';
 import 'package:sams_app/features/Grades/presentation/view/grades_tab_view.dart';
-import 'package:sams_app/features/announcements/presentation/view/announcements_tab_view.dart';
+import 'package:sams_app/features/announcements/presentation/view/announcement_tab_view/announcements_tab_view.dart';
 import 'package:sams_app/features/assignments/presentation/view/assignments_tab_view.dart';
 // Repos & Cubits
 import 'package:sams_app/features/auth/data/repos/auth_repo.dart';
